Various Tidyups, converted to 64-bit (Library DLL only)

+++ Scripts/Powershell/KodiMusicPlaylists.ps1
@@ -7,14 +7,22 @@
 .LINK  
 #>
 
-Param(
-  [Parameter(Mandatory=$True)][string]$Cred
-)
+###############################################
+# GetValues from Registry
+$LattuceRegKey=Get-ItemProperty -Path Registry::HKEY_LOCAL_MACHINE\SOFTWARE\Lattuce
 
-Add-Type –Path 'C:\Program Files (x86)\MySQL\MySQL Connector Net 6.9.8\Assemblies\v4.5\MySql.Data.dll'
+###############################################
+# Keep the master song tables in sync with the Kody databases
+[Reflection.Assembly]::LoadFile("c:\program files\Lattuce\Library.dll")
+$Utility = new-object Music.Utility
+$Utility.SyncMusicFromKodi()
+
+###############################################
+# Update the General playlists
+Add-Type –Path 'c:\program files\Lattuce\MySql.Data.dll'
 $Year=get-date -Format yyyy
 $Connection = [MySql.Data.MySqlClient.MySqlConnection]@{ConnectionString=''}
-$Connection.ConnectionString="server=lattuce-dc.lattuce.com;uid=xbmc;pwd=$Cred;database=mymusic56"
+$Connection.ConnectionString="server=$($LattuceRegKey.MySQLServer);uid=$($LattuceRegKey.MySQLUsername);pwd=$($LattuceRegKey.MySQLPassword);database=$($LattuceRegKey.KodiMusicSchema)"
 $OutputFile = "\\10.10.1.9\Userdata\playlists\music", 
               "\\media-gym.lattuce.com\c$\Users\media\AppData\Roaming\Kodi\userdata\playlists\music",
               "d:\media\music\master\playlists",
@@ -55,14 +63,25 @@ for($i = 0; $i -lt $PlayLists.Count; $i++) {
 	$NumberOfDataSets=$MYSQLDataAdapter.Fill($MYSQLDataSet, "data")
 
 	# Remove the Existing Files if they exist
-	$OutputFile | foreach {if (test-path $_\$PlayLists[$i].File) {Remove-Item $_\$PlayLists[$i].File}}
+	$OutputFile | foreach {
+		if (test-path $_\$($PlayLists[$i].File)) {
+			write-output "Deleting Playlist File: $_\$($PlayLists[$i].File)"
+			Remove-Item $_\$($PlayLists[$i].File)
+		}
+	}
 
 	# Write the New Playlist
 	foreach($DataSet in $MYSQLDataSet.tables[0]) {
 		$OutputFile | foreach {
-			$DataSet[0] | Out-File $_\$($PlayLists[$i].File) -append -encoding ASCII
+			if (test-path($_)) {
+				$DataSet[0] | Out-File $_\$($PlayLists[$i].File) -append -encoding ASCII
+			}
 		}
 	}
 }
+###############################################
+# Update the google playlists
+python ..\Python\GooglePlayPlaylists.py
 
+write-output "Done"
 $Connection.Close()
